LOOKDEVX-281 : Fix integer / string conversion in Runtime  (#908)
* Fix int conversion to string.

* Fix test.

* Revert formatting change.

* Avoid existing unstable function call.

+++ source/MaterialXRuntime/Private/PvtTypeDef.cpp
@@ -119,7 +119,7 @@ template <> void toStringValue<float>(const RtValue& src, string& dest)
 template <> void toStringValue<int>(const RtValue& src, string& dest)
 {
     std::stringstream ss;
-    ss << src.asFloat();
+    ss << src.asInt();
     dest = ss.str();
 }
 template<class T>
